requirements.txt
requests==2.31.0
google-generativeai==0.3.2
pathlib==1.0.1